requirements.txt
customtkinter>=5.2.2
pillow>=10.4.0
opencv-python>=4.9.0.80
albumentations>=1.3.1
pyyaml>=6.0.1
numpy>=1.26.0